feature: Implement Single Cycle MIPS32 Subset
