+++ asm/rom.asm
@@ -1,6 +1,6 @@
-MOV B,1
-IN A
-ADD A,1
-MOV B,A
-OUT B
-JMP 5
+start:
+    addi $v0, $zero, 1
+    addi $v0, $v0, 1
+    j end
+end:
+    j end
